Dockerfile: Dockerized build for xilinx_k8s_fpga_plugin image
Current build system is not containereized and requires you to
install golang and other tools. With docker this is not a requirement
anymore, plus you get automatization and a stable enviroment for
CI/CD. Additionally you can compile for other architectures such as
arm64, ppc...

The build process is divided in two stages so that the final deployable
image doesn't contain unnecessary files and therefore the image is
lightweight. Compare current 686.9 MB image with the new one of 44.11 MB

Signed-off-by: Pablo Rodriguez <pablo.rodriguez-quesada@windriver.com>