fpga: Add Edge devices support
Current plugin only support PCIe devices but XRT supports Zynq platform
as well.

To do so GetDevicesEdge function was created. In arm64 we have a DTB
with the device information. The program checks if the file exists and
if so we continue to the Edge custom function, if not, PCIe code is run
unmodified.

GetDevicesEdge is a simplified copy of GetDevices for PCI.
Main differences:
- As we don't have a PCI id and on Zynq ther is only one FPGA 1 is used
as the ID. This is done also in XRT, see:  https://github.com/Xilinx/XRT/
blob/2019.2/src/runtime_src/core/edge/tools/xbutil/xbutil.cpp#L402
- Shell version and timestamp does not exist in Edge either so the model
of the board is used insted. This is made so that is I want to orchestr-
ate N Edge devices and I need any FPGA I can select one from a pool
For instance: limits: xilinx.com/fpga-ZynqMP-ZCU102: 1
All this information is extracted from the device-tree.
- Management for the pairmap is left empty

Signed-off-by: Pablo Rodriguez <pablo.rodriguez-quesada@windriver.com>